+++ src/test/setup.ts
@@ -1,4 +1,14 @@
 import "@testing-library/jest-dom/vitest";
+
+Object.defineProperty(Range.prototype, "getClientRects", {
+  configurable: true,
+  value: () => [],
+});
+
+Object.defineProperty(Range.prototype, "getBoundingClientRect", {
+  configurable: true,
+  value: () => new DOMRect(),
+});
 import { cleanup } from "@testing-library/react";
 import { afterEach } from "vitest";